se realizaron cambios para poder crear una caja y poder verificar si existe, según la fecha del dia
