+++ src/database/models/Caja.js
@@ -1,19 +1,18 @@
 import mongoose from "mongoose";
 
 const cajaSchema = new mongoose.Schema({
-    ganancias: {
-        type: Number,
-        required: true,
-      },
-      cantidadPedidos: {
-        type: Number,
-        required: true,
-      },
-      fechaCierre: {
-        type: Date,
-        default: Date.now,
-      },
-})
+  ganancias: {
+    type: Number,
+    required: true,
+  },
+  cantidadPedidos: {
+    type: Number,
+    required: true,
+  },
+  fechaCierre: {
+    type: String,
+  },
+});
 
-const Caja = mongoose.model('caja', cajaSchema)
-export default Caja
+const Caja = mongoose.model("caja", cajaSchema);
+export default Caja;
